apps/worker-image-forensics/pyproject.toml
[project]
name = "worker-image-forensics"
version = "0.1.0"
description = "VIGIL APEX image-forensics worker — signature similarity, font / typography anomaly, EXIF strip. Feeds patterns P-G-002 and P-G-004."
requires-python = ">=3.12,<3.13"
license = { text = "UNLICENSED" }
authors = [{ name = "Junior Thuram Nana", email = "satoshinakamotobull@gmail.com" }]

dependencies = [
  "vigil-common @ file:///app/packages/py-common",
  "opencv-python-headless>=4.10.0,<5",
  "scikit-image>=0.24.0,<1",
  "Pillow>=10.4.0,<11",
  "numpy>=1.26.0,<3",
  "imagehash>=4.3.1,<5",
  "exifread>=3.0.0,<4",
  "ipfshttpclient>=0.7.0a1; python_version<'3.13'",
  "httpx>=0.27.0,<1",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0,<9",
  "pytest-asyncio>=0.24.0,<1",
  "pytest-cov>=5.0.0,<6",
  "pytest-mock>=3.14.0,<4",
  "fakeredis>=2.24.0,<3",
]

[project.scripts]
worker-image-forensics = "vigil_forensics.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
vigil_forensics = ["py.typed"]
